backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pinecone
cohere
google-generativeai
openai
langchain-text-splitters
tiktoken
pypdf
